src/requirements.txt
fire
requests
PySocks
urllib3
anticaptchaofficial